Updated sequential version
