+++ K-Means_Clustering/Sequential-CPU/Makefile
@@ -1,6 +1,8 @@
 				
-kmeansC		:	kmeansC.c ImageStuff.c ImageStuff.h
-				gcc kmeansC.c ImageStuff.c -o kmeansC 	
-			
-				
+sequential	:	sequential.c sequential.h ImageStuff.c ImageStuff.h
+				gcc sequential.c ImageStuff.c -o kmeansSeq	
+clean	:
+				rm -f kmeansSeq
+images_disappear	:
+				rm -f *.bmp *.png *.jpg
