--- K-Means_Clustering/Sequential-CPU/source_file
@@ -1,4 +1,0 @@
-module load cuda
-module load gcc/4.7.2
-module load openblas/0.2.15/b1
-module load opencv/2.4.8/b1
